refactor(接口测试): 优化环境获取接口模块树的参数
