+++ backend/services/api-test/src/test/java/io/metersphere/api/controller/ApiDefinitionModuleControllerTests.java
@@ -916,26 +916,26 @@ public class ApiDefinitionModuleControllerTests extends BaseTest {
         BaseTreeNode a1a1Node = getNodeByName(baseTreeNodes, "500-test-root-20");
         assert a1a1Node != null;
         ApiModuleDTO apiModuleDTO = new ApiModuleDTO();
-        apiModuleDTO.setId(a1a1Node.getId());
-        apiModuleDTO.setStatus("None");
+        apiModuleDTO.setModuleId(a1a1Node.getId());
+        apiModuleDTO.setContainChildModule(false);
         apiModuleDTOS.add(apiModuleDTO);
         BaseTreeNode a40Node = getNodeByName(baseTreeNodes, "500-test-root-40");
         assert a40Node != null;
         apiModuleDTO = new ApiModuleDTO();
-        apiModuleDTO.setId(a40Node.getId());
-        apiModuleDTO.setStatus("All");
+        apiModuleDTO.setModuleId(a40Node.getId());
+        apiModuleDTO.setContainChildModule(true);
         apiModuleDTOS.add(apiModuleDTO);
         BaseTreeNode a50Node = getNodeByName(baseTreeNodes, "500-test-root-50");
         assert a50Node != null;
         apiModuleDTO = new ApiModuleDTO();
-        apiModuleDTO.setId(a50Node.getId());
-        apiModuleDTO.setStatus("Current");
+        apiModuleDTO.setModuleId(a50Node.getId());
+        apiModuleDTO.setContainChildModule(false);
         apiModuleDTOS.add(apiModuleDTO);
         BaseTreeNode a100Node = getNodeByName(baseTreeNodes, "500-test-root-100");
         assert a100Node != null;
         apiModuleDTO = new ApiModuleDTO();
-        apiModuleDTO.setId(a100Node.getId());
-        apiModuleDTO.setStatus("All");
+        apiModuleDTO.setModuleId(a100Node.getId());
+        apiModuleDTO.setContainChildModule(true);
         apiModuleDTOS.add(apiModuleDTO);
         result = this.requestPostWithOkAndReturn("/api/definition/module/env/tree", new EnvApiModuleRequest() {{
             this.setProjectId(project.getId());
